APPC-847
-verifyPurchase method created
-Security class moved to java
-java.util.Base64 imported instead of android.util.Base64
-code adapted to new import
