+++ appcoins_billing/src/main/java/com/appcoins/sdk/billing/AppCoinsBilling.java
@@ -1,13 +1,13 @@
 package com.appcoins.sdk.billing;
 
-
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import sun.rmi.runtime.Log;
 
 public class AppCoinsBilling implements Billing {
   private final Repository repository;
   private final String base64PublicKey;
-
 
   public AppCoinsBilling(Repository repository, String base64PublicKey) {
     this.repository = repository;
@@ -16,10 +16,28 @@ public class AppCoinsBilling implements Billing {
 
   @Override public PurchasesResult queryPurchases(String skuType) {
     try {
-      //if(Security.verifyPurchase()){
+      PurchasesResult pr = repository.getPurchases(skuType);
+      ArrayList<Purchase> invalidPurchase = new ArrayList<Purchase>();
 
-      //}
-      return repository.getPurchases(skuType);
+      for (Purchase purchase : pr.getPurchases()) {
+        String purchaseData = purchase.getOriginalJson();
+        String signature = purchase.getSignature();
+
+        try {
+          if (!Security.verifyPurchase(base64PublicKey, purchaseData, signature)) {
+            invalidPurchase.add(purchase);
+          }
+        } catch (IllegalArgumentException e) {
+          invalidPurchase.add(purchase);
+        }
+      }
+
+      if (invalidPurchase.size() > 0) {
+        pr.getPurchases()
+            .removeAll(invalidPurchase);
+      }
+
+      return pr;
     } catch (ServiceConnectionException e) {
       e.printStackTrace();
       return new PurchasesResult(Collections.emptyList(), -1);
