+++ toolbox/build.gradle
@@ -8,7 +8,7 @@ android {
 
   defaultConfig {
     applicationId "com.appcoins.trivialdrivesample"
-    minSdkVersion 21
+    minSdkVersion 26
     targetSdkVersion 28
     versionCode 11
     versionName "1.3.4"
